+++ spython/tests/test_client.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# Copyright (C) 2017-2019 Vanessa Sochat.
+# Copyright (C) 2017-2020 Vanessa Sochat.
 
 # This Source Code Form is subject to the terms of the
 # Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
@@ -29,16 +29,16 @@ def test_export():
 
 def test_pull_and_run(tmp_path):
     pass
-    #image = Client.pull("shub://vsoch/singularity-images", 
-    #                    pull_folder=str(tmp_path))
-    #print(image)
-    #assert os.path.exists(image)
-    #ext = 'sif' if Client.version_info().major >= 3 else 'simg'
-    #assert image == str(tmp_path / ('singularity-images.' + ext))
+    image = Client.pull("shub://vsoch/singularity-images", 
+                        pull_folder=str(tmp_path))
+    print(image)
+    assert os.path.exists(image)
+    ext = 'sif' if Client.version_info().major >= 3 else 'simg'
+    assert image == str(tmp_path / ('singularity-images.' + ext))
 
-    #result = Client.run(image)
-    #print(result)
-    #assert 'You say please, but all I see is pizza..' in result
+    result = Client.run(image)
+    print(result)
+    assert 'You say please, but all I see is pizza..' in result
 
 def test_docker_pull(docker_container):
     tmp_path, container = docker_container