+++ spython/utils/terminal.py
@@ -1,6 +1,6 @@
 '''
 
-Copyright (C) 2017-2019 Vanessa Sochat.
+Copyright (C) 2017-2020 Vanessa Sochat.
 
 This Source Code Form is subject to the terms of the
 Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
@@ -9,6 +9,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 import os
+import pwd
 import re
 import semver
 from spython.logger import bot
@@ -81,6 +82,17 @@ def get_singularity_version():
                 version = version['message'][0].strip('\n')
 
     return version
+
+
+def get_userhome():
+    '''get the user home based on the effective uid
+    '''
+    return pwd.getpwuid(os.getuid())[5]
+
+def get_username():
+    '''get the user name based on the effective uid
+    '''
+    return pwd.getpwuid(os.getuid())[0]
 
 def get_singularity_version_info():
     '''get the full singularity client version as a semantic version"